Added comments to all handler functions
