+++ cmd/attachment.go
@@ -45,6 +45,7 @@ func (ar attachmentReader) ReadFile(path string) (io.Reader, error) {
 	return reader, nil
 }
 
+/* attachmentHandler uploads an attachment (file, image, etc) to Gitlab and returns metadata about the upload. */
 func (a *api) attachmentHandler(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json")
 	if r.Method != http.MethodPost {
