+++ cmd/job.go
@@ -15,6 +15,7 @@ type JobTraceResponse struct {
 	File string `json:"file"`
 }
 
+/* jobHandler returns a string that shows the output of a specific job run in a Gitlab pipeline */
 func (a *api) jobHandler(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json")
 	if r.Method != http.MethodGet {
