+++ cmd/members.go
@@ -12,6 +12,7 @@ type ProjectMembersResponse struct {
 	ProjectMembers []*gitlab.ProjectMember
 }
 
+/* projectMembersHandler returns all members of the current Gitlab project */
 func (a *api) projectMembersHandler(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json")
 	if r.Method != http.MethodGet {
